refactor: renames delete function from StockService.java file to deleteByEntrepreneurId; feat: getting an empty stock will now throw a ResourceNotFound exception.

+++ src/main/java/br/ifrn/edu/sisconf/service/StockService.java
@@ -28,8 +28,11 @@ public class StockService {
         stockRepository.save(stock);
     }
 
-    public void delete(Entrepreneur entrepreneur) {
-        Stock stock = stockRepository.findByEntrepreneurId(entrepreneur.getId());
+    public void deleteByEntrepreneurId(Long entrepreneurId) {
+        Stock stock = stockRepository.findByEntrepreneurId(entrepreneurId);
+        if (stock == null) {
+            throw new ResourceNotFoundException("Este estoque não existe");
+        }
         stockRepository.deleteById(stock.getId());
     }
